Create countInversion.py
In this inversion count we used the merge sort to count inversion. since an inversion in an array can be of three types, Left Inversion, right Inversion and split inversion. Inversions occurring in the left partition of the array are known as Left inversion and those occurring in right Partition are known as right inversion. while the inversion occurring in both of them i.e right and left are known as Split inversion. Total number of inversions is sum of all these inversions i.e. 
Total Number of Inversion = (Left Inversion)+(Right Inversion)+(Split Inversion)